Refactoring and simplifying slack notifications.
